(PDB-1485) Add default PuppetDB root context
This commit changes the terminus configuration to use the new root
prefix for PuppetDB and changes the connection validator to do the same.

This commit also adds the terminus package name changes `terminus` to
`termini` and tries to be smart about the default, checking to see what
version of PuppetDB was passed in and defaulting off of that value.